pyproject.toml
[tool.poetry]
name = "portfolio-backend-fastapi"
version = "0.1.0"
description = ""
authors = ["Dima <dima.horbal.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
alembic = "^1.13.1"
sqlalchemy-utils = "^0.41.2"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pydantic = "^2.7.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
